==> code/1_Routes.ipynb <==
# Table of Contents

- [Imports](#Imports-and-Configuration)
- [Routing](#Routing)
- [Save](#Save-Results)

# Imports and Configuration

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import os
import googlemaps
import requests
from googlemaps import convert
# googlemaps. configure(api_key='AIzaSyA_iFfoDyTCV-ENgl8m0FSwk1qy44EOcjw')

# Get API key
with open('../api_key_maps.txt') as f:
    api_key = f.readline()
    f.close

# Routing

To generate routes we use the google directions API.

# Get user input for departure and arrival cities
departure = input('Where are you leaving from? ')
destination = input('Where are you going? ')

# get the route from the user inputs. 
# alternatives = True tells the function to return more than 1 route (if there are any) 
# so that the user can choose the one he prefers
client = googlemaps.client.Client(key = api_key)
routes = googlemaps.directions.directions(client = client,
                                          origin = departure,
                                          destination = destination, 
                                          alternatives = True, 
                                          mode = 'driving')

# Not really an area calculation, but to filter out the results for stopover cities to only return big cities 
def areaid(place_id):
    # return the bounding box coordinates for the city
    place = googlemaps.places.place(client=client, place_id = place_id, fields=['geometry'])
    
    # assign them to variables
    max_lat = place['result']['geometry']['viewport']['northeast']['lat']
    max_lng = place['result']['geometry']['viewport']['northeast']['lng']
    min_lat = place['result']['geometry']['viewport']['southwest']['lat']
    min_lng = place['result']['geometry']['viewport']['southwest']['lng']
    
    # get the "area" and return 
    height = abs(max_lat - min_lat)
    width = abs(max_lng - min_lng)
    area = height*width
    return area

# Explore each route and look for stopover cities within the routes
client = googlemaps.client.Client(key = api_key)

# Initialize list of routes with stopover cities
cities_route = [[] for a in range(len(routes))]

# Loop through each route
# routes are broken down as: route -> legs -> steps
for index, route in enumerate(routes):
    # we always have only 1 leg here since we didnt specify any waypoints 
    for leg in route['legs']:
        # loop through each step (in this case they make up for all the route)
        for step in leg['steps']:
            # get the end coordinates
            lat = step['end_location']['lat']
            lon = step['end_location']['lng']
            # look for cities in that area (within 5 Km)
            places_found = googlemaps.places.places_nearby(client = client, 
                                                           radius = 5000,
                                                           location = (lat, lon),
                                                           type = 'cities')
            
            # go through the results and select only those that have a big enough area (using function defined above)
            for place in places_found['results']:
                if (place['name'] not in cities_route[index]):
                    if ('locality' in place['types']) and ('political' in place['types']):
                        # from observations the big cities always had that value greater than 0.1
                        if(areaid(place['place_id']) > 0.1):
                            cities_route[index].append(place['name'])

# print the alternative routes
for route_number in range(1, len(cities_route)+1):
    print("Route Number ", route_number, ':')
    print(cities_route[route_number-1]) 

# let user choose the best one
choice = ''
while (choice != '1' and choice != '2' and choice != '3'):
    choice = input('Choose your favorite route ')
chosen_route = cities_route[int(choice)-1]

# Save Results

# save chosen route to a dataframe
routes_df = pd.DataFrame({'Cities': chosen_route})
routes_df.to_csv('../datasets/route.csv')
